Move transform_output to helpers module
